Switch zaqar-ui to release:cycle-with-milestones
zaqar-ui was using (by default) an independent release model,
although it was never released. However it would prefer to use
stable branches over a forever compatible-with-all-past-releases
model, so a cycle-following model is better suited for its goals.
Since it was never released before, it is still time to adjust it
before we process newton-2.

Change-Id: I8bc491cb85025c9ba36d0f1358e86add99bd8d37