Change glance-store name to match launchpad project
When glance-store was created in launchpad [1], we could not create a
project with '_' in it's name due to launchpad constraints.

This commit modifies the reference to reflect the name of the project to
match that of the launchpad name. This is also important as the release
team refers to the project to match the launchpad name and then asks for
the repo name separately. `glance-store` launchpad project uses
openstack/glance_store repo [2] for it's development.

This governance change is thus required to unblock release of the
glance_store library.

Release team README [3] referring launchpad project name requirement as
of July 18th 2016.

[1] https://bugs.launchpad.net/glance-store
[2] http://git.openstack.org/cgit/openstack/glance_store/
[3] http://git.openstack.org/cgit/openstack/releases/tree/README.rst#n153

Change-Id: Idd3f57672466b894d4f33be1a22a4e725e562970