Storlets to become official - Proposed governance change
[1] Has a summary as well detailed desciption of all that
the team has done following the comments given in August.

The below is the original commit message describing the project
(with some necessary updates).

The project offers the ability to run user defined functions
within Openstack Swift nodes. For more info please refer to
the docs (built from source) here:
http://storlets.readthedocs.io/en/latest/

The project combines a Swift middleware together with Docker
to execute the user defined Java/Python programs inside a container,
which in turn runs on Swift nodes. In addition the repository
has several sample storlets used for functional tests
(as well as examples).

Open Source:
License: Apache 2.0 license.
Libraries: On top of what swift requires, storlets make use
of docker, OpenJDK JRE-8 plus standard Java libs. Non standard java
libs that we use are:
- simple-json (Apache 2.0 lic)
- logback-classic, logback-core (EPL v1.0 and the LGPL 2.1)
- slf4j-api (MIT lic)

Open Community:
We are currently 5 active contributors, 1 from IBM, 3 from NTT
and this commit author (independent developer) we hold weekly
meetings in #openstack-meetings, with agenda posted in [2]

Open Development:
The project is integrated with the Openstack CI, and was created
according to Openstack "Project Creator's Guide".
We work closely with the Swift team, and one of our cores is also a
core in Swift. Yours truly was elected to be the first PTL by the team.
Our current and near term goals are documented in our wiki.

Open Design:
So far our design meetings were alongside those of Swift. Otherwise,
most of the work is done through Gerrit and our IRC channel. In addition
we have the [storlets] prefix within the openstack-dev ML.

Interoperability:
Our user facing APIs are actually an extension of Swift’s API
(http://storlets.readthedocs.io/en/latest/api/overview_api.html)
To the best of my understanding, storlets being a Swift extension
do not pose any authentication restriction or integration issues.

Testing interface:
We have regular unit/functional/python style tests running with every
review request. We generate a source tar ball of the python code as
part of the deployment scripts, and generate documentation.

The project has been presented in the Paris, Tokyo and Barcelona
summits.

Reasoning about languages used:
We use Python, Java and C.
Java and Python are used as the programming language for storlets.
That is, the user defined code running inside the Docker container on the storage
node can be Java or Python. Java is widly used in the Big-Data analytics world as
well as in the media world (as demonstrated in our Paris and Barcelona talks).
Similar to OpenStack SDK we would like to support a set of bindings and we are
currently working on a Python 'binding'.

We use C  for passing fds between Swift and the Dockered storlet, and only for that
purpose. Unfortunately, neither python 2.7 nor Java have support of passing fds
between processes (over UDS). This feature is crucial for us as we want to keep the
Docker container isolated as much as possible. Apparently, python 3.x does have
such support, but seems that on the Java side we will still need the C code.

Python is used for the Swift middleware. The middleware intercepts Swift requests
that wish to invoke a storlet on some data and routes the data to/from the storlet.

JRE Dependency:
We are now using OpenJDK.

Is this being implemented anywhere else:
There have been https://zerovm.readthedocs.io/zerocloud/overview.html
However, after being aquired by Rackspace they have 'disappeared'.
We have tried to design storlets so that different runtime sandboxes
are pluggable with Docker containers being our first implementation
of a runtime sandbox. Having said that, as long as we have only one
implementation, this design has not been put to the test.

There is also what has been a proprietary solution and now opensourced
https://www.joyent.com/manta
Joyent was aquired by Samsung this year.

[1] https://etherpad.openstack.org/p/storlets-big-tent
[2] https://wiki.openstack.org/wiki/Meetings/Storlets

Change-Id: Icf9dd346c731dd7378ba9015bd3cbc31e4246a6a