Move puppet-tripleo to TripleO
puppet-tripleo was owned by Puppet OpenStack and TripleO teams but the
actual users and developers are TripleO folks.

In term of ownership it would make sense to move it into TripleO
governance.
It will mean that it will be released in same time as other TripleO
projects and also core reviewed by TripleO developers.

Change-Id: I4dbf15dc32b27ab8e54058445384887e4a97e0ba