Add tag for devstack plugins
This commit adds a new deliverable type tag to describe devstack
plugins. For right now this is just for standalone plugin repos.
However the vast majority of devstack plugins are packaged in the
project repos. It might be worth adding the tag to those repos as
well. But that is a separate discussion to have and will happen in,
a follow one change.

Change-Id: Ia4484793a11810a0835d12f1ea7e1e1bb3d66e18