Create a project tag for zero-downtime upgrades
In transitioning projects towards a rolling upgrades strategy, it's
become clear that there are additional levels of granularity, reflecting
the additional difficulty of implementing and maintaining acceptable
solutions.

The existing asserts:supports-rolling-upgrade tag allows for minimal
downtime of the control plane during the upgrade process. This tag
raises the bar further by eliminating that allowance.

As a concession, this tag still allows for performance degradation
during the upgrade process, which we could again eliminate in a more
restrictive tag as projects begin to assert this one.

With this patch, the upgrade-related tags would then include:

1. asserts:supports-upgrade
2. asserts:supports-rolling-upgrade
3. asserts:supports-zero-downtime-upgrade

Change-Id: I8b191f7d56702aa21273c61417af4bdf1de2abb0