Adjust TC and PTL election timeframes
With the Design Summit functions being split between the OpenStack
Summit and the Project Teams Gathering, the current wording in the
TC charter around election (which uses "Design Summit" and "Summit"
interchangeably) is no longer valid.

This change clarifies the situation by:
* keeping PTL elections near to development cycle boundaries
  (but defining the timeframe relative to release rather than
  relative to an event)
* keeping TC elections on a 6-month spring/fall cadence,
  organizing them relative to the OpenStack Summit dates

The change is compatible with the language around TC elections
present in the OpenStack Foundation bylaws (esp. Appendix 4),
which mentions 6-month and one-year periods but does not attach
TC elections to events or cycles. This change therefore should
not require any change on the bylaws side.

While it tries to minimize impact, the net result of the change
is that TC elections are now separate from PTL elections (rather
than run just the week after).

This change also adds some flexibility in vote organization, by:

* using "on or before" language to allow election officials to
  work around holidays
* removing language around nomination periods (potentially allowing
  for the addition of "campaign days" between nomination and vote)
* lowering the minimum required amount of voting days to 4 business
  days (potentially allowing the vote to be organized within a week
  rather than forcing it to overlap with a weekend)

Change-Id: I5900f9b4a10cbdb09609516d0b04f29b4fe7eb88