Move incomplete OpenStack-Ansible projects to release:none
The following OpenStack-Ansible repositories that are currently listed
as release:cycle-trailing are not ready for a release cycle, and are
currently in early development or non-functional:

os_barbican
os_cloudkitty
os_designate
os_monasca
os_watcher
os_zaqar

This PR moves these repositories to release:none. These will be moved to
follow a release once these repositories are functional and ready.

Change-Id: I1996302e78a6bda934e67b13e65941e500ecb5e2