remove release team tags
Release models for deliverables may change over time, so we are now
tracking them in the releases repository, where we have a separate file
for each release cycle.

The deliverable "type" tags were only being used to control how the
releases.openstack.org site organizes the data about releases on each
series page, so we are also now tracking the type in the releases
repository.

Since the data associated with these tags is tracked elsewhere, and not
used for governance directly, we can remove the tags.

Change-Id: Iac89ba587ca770004264aaebf1bc856b755e2258
Signed-off-by: Doug Hellmann <doug@doughellmann.com>