Get rid of the trove-image-builder repository
The project was to store image elements for Trove. These elements are
now in the trove repository and therefore branched with the project,
used in the CI, and so on. In short, we don't need this independent
repository any longer.

Change-Id: I4a9381ab9a11b49fb4a8c227a39a5971a0c5cccb
Related: I4186340ef13cdbaa2df0a752b68e2e4bd6968494