Add project Octavia to OpenStack big-tent
This involves merging neutron-lbaas into the Octavia
project and migrating Octavia out of the neutron stadium, making
Octavia the Load Balancing As A Service (LBaaS) project for
OpenStack. Specs for exactly how this is going to happen are
linked as depends-on. The 'neutron-lbaas' repo will exist during
the transition, and eventually be retired with no impact on
operators/end users beyond packaging.

Change-Id: Ia807a5ad36859a0228aed76c307a97dec8494b7c
Depends-On: I5303b04ec2ca5acd525d9b5137f012515a8dc1df
Depends-On: I9bc348c6f86f868ec5e9ab2be2ae49947bfbacdc