Disallow downtime of controlled resources during upgrades
One of the outcomes of the cross-project upgrades session in Barcelona
was that upgrades should not incur *any* downtime of resources
controlled by OpenStack [1]. To paraphrase someone in that session who
explained the reasoning more elegantly than I ever could:

    If you expect my virtual machine to go down during an upgrade, then
    I might as well migrate my workload into a different cloud. That's
    not an upgrade at all.

The basic upgrade tag now includes an amendment to more accurately
reflect that tag's original intent: controlled resources should not
experience any change of state due to upgrades (for example, virtual
machines should not require rebooting just because nova is being
upgraded from Newton to Ocata).

In order to avoid changing the requirements of the existing basic
upgrade tag, a new tag is introduced which asserts that the
accessibility of controlled resources must be maintained during the
upgrade (for example, an SSH session to a virtual machine should not be
interrupted just because nova is being upgraded from Newton to Ocata).

The zero-downtime upgrade tag continues to focus on control plane
downtime (downtime of the service APIs themselves).

[1] "Data plane vs control plane" (~L57) https://etherpad.openstack.org/p/ocata-xp-upgrades

Change-Id: Ia3ebbce4ac18ccfbd4167ef519b490b95886333d