Generalize tag application processes
Prior to this change, each tag had it's own specific tag application
process, sometimes with zero (or just inconsequential) changes in
wording when compared to a similar tag.

In order to eliminate these redundancies, this patch proposes to
generalize the tag application process on the main tags page, and allow
each tag, or group of tags, to further specify additional process
requirements (such as requiring prior approval by a group other than the
TC, or identifying the group expected to submit tag applications).

This should also make it easier to review new tags in the future.

Change-Id: I7d78dcc95ed51794c2456c827187a7087126dd21