Document current base services
Introduce the concept of "base services" (external services that
OpenStack components can assume will be present in any OpenStack
installation) and list current base services.

This is intended to reflect the current situation. Changes to the
list should be proposed as subsequent patches.

You can find a more detailed analysis and rationale for this change
in the Architecture WG analysis:

http://git.openstack.org/cgit/openstack/arch-wg/tree/active/base-services.rst

Change-Id: I77ffc000d895c3d8eb27b3b821c6c63f9dd2f675