Move shade into its own top-level team
shade started as a utility library for Infra, but has grown into a
popular and widely-used client library in its own right. However,
even though it's a solidly maintained library and is licensed 
appropriately, it is not seen as an OpenStack deliverable.
While there is no rule that says that OpenStack must produce all of the
software it depends on, and in fact, OpenStack itself consumes a vast 
amount of dependencies produced elsewhere, the fact that shade is 
specifically focused on the OpenStack End User and the OpenStack APIs
makes it seem like it should be an official output of the OpenStack 
project. 

This is the first instance of moving a project from Infra to OpenStack,
so there are a few weirdnesses - notably shade publishes docs to
docs.openstack.org/infra not docs.openstack.org/developer and the shade
repo is in openstack-infra/shade not openstack/shade. We're not
suggesting changing the git repo due to it being unneccesary disruption.
The docs can be moved though, as it's easy to host a redirect.

Additionally, from shade's inception we've held that shade's existence
is a bug. However, as much fun as that is to say at conferences, perhaps
the time has come to put the negativity of that trope to rest and
instead focus on continuing to empower our users. It'll likely be a few
years yet before we could 'fix' all of the APIs to make shade unneeded
and even then we'll still have old clouds for a few more years that users
will still need to interact with. So as a general multi-cloud aware library
it'll likely be around for some time and should go ahead and be cared
for and fed.

While we're at it, move requestsexception over too. It was written for
and is used by shade, so having it be part of the same deliverable chain
is preferrable.

Change-Id: I7de2937bc18868c1c0c5c4dc2efbe3ede5865494
Signed-off-by: Monty Taylor <mordred@inaugust.com>