update sphinx and treat warnings as errors
Sphinx 1.5.0 and pbr 2.0 include a feature to let us pass the flag to
treat warnings as errors through setup.cfg. Update the version of Sphinx
we're using explicitly, and rely on an unbounded pbr specification to
give us the latest version on every run.

Also fix the one warning that caused the build to fail.

Change-Id: Ie93a3adf82b8f568a3cc4357c1e78da22ddc92c0
Signed-off-by: Doug Hellmann <doug@doughellmann.com>