Remove Gnocchi from Telemetry
After a lot of talk within the Gnocchi team, it appeared to us that Gnocchi,
which has been wisely tagged as 'independent' since its inception, has a lot of
potential usage outside of OpenStack directly.

Being part of the big tent helped the project to be built, but it now appears
that it restrains its adoption and contribution from users outside of the
OpenStack realm.

Therefore this patch un-lists Gnocchi from the OpenStack Telemetry project.

Change-Id: I4979e2186e4c3a84a3c205d69a288443c3e7e463
Depends-On: Ie8c58659bc5ae820bda27414e916b41e38a62f46