Include instructions on how to propose changes
The instructions on how to propose governance changes in
OpenStack are currently located on the TC meeting agenda
wiki page, which makes it hard to discover when you browse
the governance website.

Place it right and center on governance.o.o so that it's
clear how to make changes to it.

Change-Id: Idb5a38fe6b503d8173178e956c95a00bb41272dd