Remove unused Debian repositories
A couple of Debian Tripleo and Oslo repos are not used at all,
there's e.g. deb-os-apply-config and deb-python-os-apply-config
and only the later is in use.

These repositories have been created wrongly, so let's remove them.

Note that the repos are not added to legacy.yaml, since the content is
mirroring official repos and these repos are unused for some time.

Change-Id: I2fa939ebcc1cb830dc3c92e5ac785b63d8ff5342
Depends-On: I1ac1a06931c8b6dd7c2e73620a0302c29e605f03