Add paunch to tripleo governance list
paunch [1] is a python library and CLI tool which forklifts the logic
of the docker-cmd heat-agents hook[2] into its own project.

The YAML format used by paunch is currently driven by the needs of
TripleO and is derived from the docker compose v1 format.

The intent to bring paunch into OpenStack was announced a week ago on
openstack-dev[3].

[1] https://github.com/steveb/paunch
[2] https://github.com/openstack/heat-agents/tree/master/heat-config-docker-cmd
[3] http://lists.openstack.org/pipermail/openstack-dev/2017-April/115405.html

Change-Id: Ib7f7f7492c06a17ccddf5dcf7efab75fa3cb3d51
Depends-On: Ia7be8a5f43033f4a1054ee9e960d3d5bf26a939a