Establish a #openstack-tc water cooler channel
In the Boston Summit, during the last forum session, we talked
about using #openstack-tc channel for socialization as opposed
to the current channel (#openstack-dev) as we may end up overloading
that channel and a separate channel is easier to keep an eye on.

We should also use this channel for the office hours.

Change-Id: I9430273aac72441dd843bc9f9178f4a217685f82