Move Fuel from official to unofficial (hosted)
This change moves Fuel out of the official OpenStack project list
(governed by the Technical Committee) and makes Fuel an unofficial
project, hosted under OpenStack project development infrastructure.

Fuel originated at Mirantis as their OpenStack installer. It was
proposed as an official OpenStack project in July 2015 and approved
in November 2015. The promise at that time was that making it
official would drive other organizations to participate in its
development and turn it into the one generic OpenStack installer
that everyone wanted. Fuel was not a small endeavor: in Mitaka and
Newton it represented more commits than Nova.

The Fuel team fully embraced open collaboration, but nevertheless
failed to attract other organizations. Mitaka and Newton were still
96% the work of Mirantis. Since October 2016, Fuel's activity has
dropped, following the gradual disengagement of its main sponsor.
Comparing activity in the 5 first months of the year, there was a
68% drop between 2016 and 2017, the largest of any official OpenStack
project. The Fuel team hasn't met on IRC for the last 3 months.
Activity dropped from ~990 commits/month (Apr 2016, Aug 2016) to
52 commits in April 2017 and 25 commits in May 2017. And there are
unsolved issues around licensing that have been lingering for the
last 6 months. Fuel was always a bit of an outlier on the OpenStack
project map, due to its scope (deploying OpenStack) but also due to
its own policies in terms of releasing and stable branches.

Despite the efforts of the Fuel team, Fuel did not become what we
hoped when we made it official: a universal installer produced by
the OpenStack community, and used across the board. It was worth a
try, we had to try, but at this point it's time to stop considering
it a part of "OpenStack" proper and make it an unofficial project
(hosted on OpenStack project development infrastructure).

Change-Id: Ie7da490ef27dbec38dc9f16564fe576f9680c226