Begin integrating vision feedback and editing for style
This patch, a followup to Ifde9f19cfa07c62320e2ce09f6d5e803357aa0b4,
attempts to do two main things:

* react to some of the feedback provided on the original document
  and in the related survey and summit sessions

* adjust the writing to be somewhat more structured (work in
  progress) and tighter

Notable changes include:

* adding section headers (current wording stubs for review)

* making room for a preamble which sets the stage for what the
  document is; many original readers did not understand the "this is
  a blog posting from the future" format

* making room for a bullet list of summary

* shrinking, slightly, the constellation section

* Being more consistent throughout about case and using the active
  voice. This is a bit difficult given the future-but-looking-back-
  format. It's not yet perfect but I hope at least improved. The
  change in voice probably has the most significant impact in terms
  of causing a large diff.

* Removing superflous "very" and "really" style words where they
  were superfluous. More of this could be done, but it might remove
  the somewhat folksy tone which appears to be there on purpose.

* Things like "drive-by" having been changed to use phrasing that
  hopefully has less risk of offending.

* Entries which felt too ambiguous or didn't add much value (such as
  the git-bot) have been removed or de-emphasized.

From my standpoint this continue to be a bit too long and too
story-like. It is not yet well suited for being read on the web.
Clearly, early drafts should be story like because that greatly
helps the authors to flesh out their ideas. But as an artifact to be
read in a useful fashion by others, that may be ideal. We probably
want to find a sweet spot in the middle.

This should be considered a work in progress both in terms of the
language and structure and the actual content. Not all feedback has
yet been incorporated.

Change-Id: I7374ccaf1c992cad0f1b1e7a0eacd17f3badd75b