Adopt python-openstacksdk into shade project
While shade has been focused on a high-level python interface to
OpenStack that hides vendor differences, python-openstacksdk has been
focused on exposing the REST APIs as they are in python. The two tasks
share enough underpinnings that having two separate teams work on two
separate efforts is not an efficient use of resources these days.

python-openstacksdk by nature of its name sounds to people like it's an
official project but has never been made one which causes confusion.

As a result of shade and openstacksdk having been parallel efforts and
the different teams having slightly divergent points of view on how to
structure abstractions, openstacksdk is sadly unusable by shade in its
current form. Such divergence at this point doesn't improve the world for
anyone.

Bring the python-openstacksdk repo officially into the shade project.
The intent will be to evolve both projects into sharing a holistic
approach to consuming OpenStack APIs with openstacksdk as the lower
level and shade as a business-logic layer on top of it. This will mean
an amount of breaking changes to openstacksdk, although as it has not
yet released a 1.0 that is less bad than it otherwise could be. As
python-openstackclient uses python-openstacksdk for neutron
interactions, great care must be taken to not break
python-openstackclient in the process.

The existing shade-core will be merged with members of the
python-openstacksdk-core team who desire to continue.

Discussions of merging or consolidating repositories will follow at a
later date if they prove to be valuable but are not being proposed at
this time.

Change-Id: Ibc0e6200121a536dc7072af9049a68eac8bb28e9