Amend leaderless program resolution timeframe
The current language puts undue pressure on election officials to
immediately notify the TC in the event of a leaderless project, while
also not providing a maximum timeframe for notification, which may leave
a project leaderless for some time in exceptional circumstances.

The replacement language places a clear bound on the notification period,
allowing election officials to take care with their reporting, and the
technical committee to have information in a timeframe that allows a new
PTL to be selected to start with their peers.

Change-Id: Id82afaf24b6fd8f8db90df635debc38a04627791