Appoint Kota TSUYUZAKI as Storlets PTL for Queens
Nobody self-nominated in time to become the Storlets PTL for the
Queens cycle, therefore the TC must appoint someone or disband
the project team.

The Pike PTL, Eran Rom, has moved to another organization and
can no longer dedicate time to Storlets. Kato Tsuyuzaki has been
the leading contributor for Storlets in the Pike cycle. He is
willing and able to fill the PTL position for the Queens cycle.

Change-Id: I4f1d0277ed34a94965e72a038aff6c84a14a0de8