Masakari - Instances High Availability Service
Mission: Provide instances high availability service for OpenStack
clouds by automatically recover the instances from failures. We are
mainly focus on solve the problems addressed in High Availability for
Virtual Machines Development Proposal [1].

Project History: Engineers from NTT started this project on github [2]
in Jul 2015 under Apache License 2.0. Newton design Summit in Austin,
Masakari team and OpenStack HA team discussed how to implement this
feature in OpenStack [3]. As Masakari team agreed to re-architecte
masakari, we started a new project openstack/masakari [4] and
re-architect and rebuild the [2] with OpenStack standards. We
continually discussed this topic in OpenStack HA team [5] and we present
our plane to moving into a converged upstream solution for-instances
high-availability in Boston Summit [6]. Masakari is an essential
component of this solution.

Satisfaction of new projects requirements:
Masakari projects follow the 4 opens,

* Open Source: All Masakari source codes, specifications, and
  documentations are use Apache License 2.0. All library dependencies
  allow for unrestricted distribution and deployment.

* Open Community: We holds weekly IRC meetings in #openstack-meeting and
  weekly agenda, meeting time, and link for previous meeting logs are in
  our wiki page[7]. We are also available in #openstack-masakari IRC
  channel for daily discussion. All the channels are open for anyone to
  interact with Masakari team.

* Open Development: Masakari uses Gerrit for code reviews and Launchpad
  for bugs and blueprints tracking [8]. Project core team checks all the
  changes before merging them. The project is integrated with the
  Openstack CI, and runs pep8, py27, and py35 jobs. Masakari is
  implemented using libraries and technologies adopted by other existing
  OpenStack projects.

* Open Design: The Masakari team has been publicly discussed project
  direction, issues, work items and priorities in etherpad [9-11] and in
  the weekly IRC meetings. We use masakari-specs [12] to discuss design
  specifications for all Masakari projects. We also use openstack-dev ML
  with prefix [masakari] for project discussion.

Masakari is compatible with OpenStack APIs, and uses Keystone middleware
to integrate with OpenStack Identity.

The Masakari team is happy to participate in any goals specified by the
TC, and meet any policies that the TC requires all projects to meet.

[1] http://specs.openstack.org/openstack/openstack-user-stories/user-stories/proposed/ha_vm.html
[2] https://github.com/ntt-sic/masakari
[3] https://etherpad.openstack.org/p/newton-instance-ha
[4] https://github.com/openstack/masakari
[5] https://wiki.openstack.org/wiki/Meetings/HATeamMeeting
[6] https://www.openstack.org/videos/boston-2017/high-availability-for-instances-moving-to-a-converged-upstream-solution
[7] https://wiki.openstack.org/wiki/Meetings/Masakari
[8] https://wiki.openstack.org/wiki/Masakari#Code_and_Bug.2FFeature_trackers
[9] https://etherpad.openstack.org/p/ocata-priorities-masakari
[10] https://etherpad.openstack.org/p/masakari-pike-workitems
[11] https://etherpad.openstack.org/p/masakari-queens-workitems
[12] https://review.openstack.org/#/q/project:openstack/masakari-specs

Change-Id: I99f7bfbf785f79d235a51ca8acc8810c4805694c