Put searchlight into maintenance mode
Searchlight is a mature technology.  The current project team still
maintains it and it is used by members of the team.
However, there is no longer anyone working full time on Searchlight.

That being said, expansion of Searchlight relies on plugins for those
projects wishing to be indexed by the service.  The design and
architecture of plugins is well-documented and the plugins can be
stored in the appropriate project's repository.  Thus development of
Searchlight's functionality can continue even while the project itself
is in maintenance mode.

The project has a current PTL, so the liason requirements for the tag
are met.

Change-Id: Iefb735d3e3b651a9f9d2231dd730ca2bddc296ea