Explicitly say which release goals were completed
During the PTG, it was mentioned that it would be useful for the goal
document to call out when a specific project completed the goal. For
example, nova completed moving policy into code during the Newton
release. Keystone completed the goal during the Pike cycle. Having
this next to the completion and planning documents makes it easier
for projects consuming this work to know what projects are done and
when they were done with the work.

Change-Id: I6b8f83ab2a1fd9d079a8493cbf6dbc47d944e0eb