Add stestr to python unit test runner cti
The testrepository project is mostly inactive and not really
maintained anymore. To fix a lot of open issues with testrepository
the stestr [1][2] project was started to fix these and actively
maintain the project. As we've started migrating things to using it
we should reflect that this is the future in the CTI for python. This
commit does that and says projects can use stestr or testrepsitory and
that the use of testr is deprecated as we migrate to using stestr
everywhere.

[1] https://github.com/mtreinish/stestr
[2] http://stestr.readthedocs.io/en/latest/

Change-Id: I5e523cfdf4f068efdc102c6a3ee18847429e28a8