Align CTI to PTI
When we talk about the PTI, we tend to say "PTI". In fact, the original
filename is "reference/project-testing-interface.rst". However, the
pre-language files use the cti acronym, as do all of the title lines.

Let's align with the acronym people use when discussing the concept.

Also, remove the suffix from the language files - they're in a directory
called "pti" already, they don't need a _pti suffix.

Change-Id: I53bc82eb2cb2125d7446a5f7d0be622118553732