Remove stable:follows-policy tag from TripleO
TL;DR stable:follows-policy tag does't fit with TripleO, it's not
helping us so let's remove it.

More context now:

We applied for the stable:follows-policy tag back in Newton but it was
rejected because we didn't follow all requirements at that time.
It was re-proposed in Ocata and approved by stable-maint team.

We made our best to not backport features during Ocata and Pike but
unfortunatly we had inputs from external teams like vendors (Dell,
Nuage, Cisco, etc) who push changes in stable branches to support their
backends, quite often late in the cycles and ask for backports.
The reason of asking for backports is because most of our Users
deploy OpenStack Newton and very rarely Ocata (and not Pike yet),
and we want to satisfy their needs to make our users happy.

This doesn't fit with stable:follows-policy and the process too heavy
for us. We will continue to do our best to not backporting features
and make stable branches as stable as we can, make our best to maintain
backward compatibility and keep our users happy. But sometimes we have to
backport some patches that are really needed on the field, by production
environments.

Now someone could ask "why not carrying patches downstream-only and
don't backport upstream?". We made huge organizational changes in our
way to work together and to ship our project. We want all our
codebase upstream first and we commit to release early and often. We
also avoid the pain of maintaining downstream repositories with
cherry-picks that are not in sync with upstream (it causes issue when
rebasing sometimes).

By leaving stable:follows-policy tag, we will probably:

- Keep being conservative in our backports but make some exceptions when
  needed (exceptions that aren't working for stable:follows-policy).
- Do not EOL Newton next month as it's still our most-popular version
  right now and we might have more backports ongoing.
  We'll probably wait until Queens final release to EOL Newton, as we
  expect Queens to be our next LTS.

Note: leaving the tag for now doesn't mean we won't re-apply again later
but in our current model where most of our users are deploying "old"
stable releases, we just don't want this tag anymore. My personal hope
is we can re-apply one day as we have our users deploying the latest
stable and being able to upgrade every 6 months, but that's another
story.

Change-Id: I3ab678f333382e67993cd1057cc5ed0a132f5224